Get rid of custom merge_sort
way slower and more allocs than the `quicksort!(reduce(vcat, vecs))` way.

from profview, lotsa time was spent on
`for j in active: curval[j] < smalest?`
So an optimization could be to use a priorityqueue here. When updating the pointer of the vec with the smallest el, put the new val in a pq.